+++ CHANGELOG.md
@@ -7,9 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Added the new class `Skipgrams` (which `Ngrams` subclasses from now).
+
 ### Changed
+- Changed package name from `ngrams` to `nskipgrams`.
+- In various methods of `Ngrams`, the argument name `order` has been changed
+  into `n`.
+
 ### Deprecated
 ### Removed
+- Removed the `total_count` method of `Ngrams`.
+  Now that `Skipgrams` is added,
+  `total_count` would be too confusing and/or not of much use in practice.
+
 ### Fixed
 ### Security